fixing bugs in new PSD features

+++ moses/src/Util.h
@@ -58,7 +58,7 @@ namespace Moses
 #define VERBOSE(level,str) { if (StaticData::Instance().GetVerboseLevel() >= level) { TRACE_ERR(str); } }
 #define IFVERBOSE(level) if (StaticData::Instance().GetVerboseLevel() >= level)
 
-const size_t EPS = 0.0001; // for floating point comparisons
+const float EPS = 0.0001; // for floating point comparisons
 inline bool Equals(float a, float b)
 {
   return fabs(a - b) < EPS;
